refactor(Sockets): create sockets index file
A `sockets/index.js` file has been created to centralise all of the
socket related exports.

The "route subscriber" which is responsible for automatically connecting
to the stream's channel when the route changes has been moved into its
own file `sockets/subscribe.js`
