+++ apps/juicebox_web/assets/js/sockets/actions.js
