+++ apps/juicebox_web/assets/js/sockets/index.js
@@ -1,0 +1,3 @@
+export { default as createSocket } from './middleware';
+export { default as subscribeToStream } from './subscribe';
+export { CONNECT_TO_CHANNEL, connectToChannel } from './actions';
