+++ apps/juicebox_web/assets/js/sockets/middleware.js
@@ -1,8 +1,8 @@
 import { Socket } from 'phoenixjs';
 import { defaults, omit, forEach } from 'lodash';
-import { CONNECT_TO_CHANNEL } from '../sockets/actions';
+import { CONNECT_TO_CHANNEL } from '../sockets';
 
-const socket = (function() {
+const newSocket = (function() {
   let socket;
 
   return function() {
@@ -15,20 +15,14 @@ const socket = (function() {
   }
 }());
 
-function connect(streamId) {
+function connect(socket, streamId) {
   const topic = `stream:${streamId}`;
-  const channel = socket().channel(topic);
-
+  const channel = socket.channel(topic);
   channel.join()
-    .receive('ok', resp => {
-      console.log('Joined successfully', streamId, resp);
-    })
-    .receive('error', resp => { console.log('Unable to join', resp) });
-
   return channel;
 }
 
-function createSocket() {
+function createSocket(socket = newSocket()) {
   return ( { dispatch } ) => {
     let channel;
 
@@ -38,7 +32,7 @@ function createSocket() {
       switch(action.type) {
         case CONNECT_TO_CHANNEL:
           channel && channel.leave();
-          channel = connect(action.channel);
+          channel = connect(socket, action.channel);
           channel.on('remote.action', action => dispatch(action));
         default:
           if (socketData) {
